Country codes

A reference to know the country codes to be able to download and import data (e.g. chl stands for "Chile" in the OEC's API). This follows the standard ISO-3 country codes.

country_codes

Format

A data frame with 263 observations on the following 2 variables.

country

Official countries' names spelled according to the United Nations.

country_code

Three characters codes to refer to the countries in the OEC's API.

Examples

country_codes

#> # A tibble: 264 x 2
#>    country             country_code
#>    <chr>               <chr>       
#>  1 Afghanistan         afg         
#>  2 Albania             alb         
#>  3 Algeria             dza         
#>  4 American Samoa      asm         
#>  5 Andorra             and         
#>  6 Angola              ago         
#>  7 Anguilla            aia         
#>  8 Antarctica          ata         
#>  9 Antigua and Barbuda atg         
#> 10 Areas               xxa         
#> # ... with 254 more rows