Convenient wrapper for getdata

Returns a tibble containing imports, exports and relevant metrics (e.g. exports growth w/r to last year) for a range of years as the result of different API calls and data transforming. The API returns data for different trade classifications: (i) SITC (1962-2016); (ii) HS92 (1992-2016); (iii) HS96 (1996-2016); (iv) HS02 (2002-2016); (v) HS07 (2007-2016).

getdata_batch(origin, destination, initial_year, final_year, classification, by)

Arguments

origin
ISO code for country of origin (e.g. chl for Chile). Run country_codes in case of doubt.

destination
ISO code for country of destination (e.g. chn for China). Run country_codes in case of doubt.

initial_year
Numeric value greater or equal to 1962 and lower of equal to 2016. This needs to be lower or equal to final_year.

final_year
Numeric value greater or equal to 1962 and lower of equal to 2016. This needs to be greater or equal to initial_year.

classification
Any of the available trade classifications in the OEC (sitc, hs92, hs96, hs02 or hs07). Default set to sitc.

by
Numeric value to define the increment of the sequence of years. Default set to 1.

See also

Examples

# NOT RUN {
# The next examples can take more than 5 seconds to compute, and specially for large economies so 
# these are just shown without evaluation according to CRAN rules

# Run `country_codes` to display the full table of countries

# What does Chile export to China?  
# years 2010 to 2015 - SITC (4 characters)
getdata_batch("chl", "chn", 2010, 2015)
# or with explicit parameters
getdata_batch("chl", "chn", 2010, 2015, "sitc", 1)

# What does Chile export to China?  
# years 2010 to 2015 - HS07 (4 and 6 characters)
getdata_batch("chl", "chn", 2010, 2015, "hs07")

# What does Chile export to China?  
# years 2010 and 2012 - HS07 (4 and 6 characters)
getdata_batch("chl", "chn", 2010, 2012, "hs07", 2)
# }